+++ src/components/Canvas/index.tsx
@@ -3,41 +3,34 @@ import styled from '@emotion/styled';
 import { useContainer } from '@pixore/subdivide';
 import { useSprite, useSpriteActions } from '../../contexts/Sprite';
 import { useArtboard, useArtboardActions } from '../../contexts/Artboard';
-import useCanvas from './useCanvas';
 import useTool from './useTool';
-import FrameLayers from './FrameLayers';
-import Background from './Background';
-import Mask from './Mask';
+import FrameLayers from '../CanvasLayers/FrameLayers';
+import Background from '../CanvasLayers/Background';
+import Mask from '../CanvasLayers/Mask';
 import { Context as ListenerContext } from '../../tools';
-import { useCanvas2DContext } from '../../hooks/useCanvas';
+import { useCanvas2DContext, useCanvas } from '../../hooks/useCanvas';
 import CanvasLayer from '../CanvasLayer';
 import PanelSelect from '../PanelSelect';
+import { HeadlessPanel } from '../Panel';
+import { usePlayAndPause } from '../../hooks/usePlayAndPause';
+import Frames from '../CanvasLayers/Frames';
+import Tools from './Tools';
+import CenterButton from '../CenterButton';
 
 const Float = styled.div`
   display: inline-block;
   position: absolute;
-  top: 0;
-  left: 0;
+  top: 4px;
+  left: 70px;
   padding: 2px;
-  border-radius: 6px;
+  border-radius: 3px;
   background: rgba(0, 0, 0, 0.5);
 `;
 
-const Container = styled.div`
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  border: 2px solid #2e3440;
-  border-radius: 10px;
-  background: #3b4252;
-  overflow: hidden;
-`;
-
 const Canvas: React.FC = () => {
+  const container = useContainer();
   const sprite = useSprite();
-  const { stats } = useContainer();
+  const { stats } = container;
   const { onRef: setMainRef, context: mainContext } = useCanvas2DContext();
   const {
     onRef: setPreviewRef,
@@ -50,6 +43,10 @@ const Canvas: React.FC = () => {
   const { onWheel } = canvas;
   const { layer, tool: toolName } = artboard;
   const { layers } = sprite;
+  const { isPlaying, button } = usePlayAndPause(
+    false,
+    sprite.frames.length === 1,
+  );
 
   const listenerContextRef = React.useRef<ListenerContext>({
     mainContext,
@@ -79,27 +76,52 @@ const Canvas: React.FC = () => {
     }
   }, [stats, canvas, sprite]);
 
+  const onCenter = () => {
+    canvas.center(stats, sprite);
+  };
+
   const indexOfCurrentLayer = layers.indexOf(layer);
   const layersBelow = layers.slice(0, indexOfCurrentLayer);
   const layersAbove = layers.slice(indexOfCurrentLayer + 1, layers.length);
 
+  // TODO while zooming in/out if the users try to paint
+  // an error happen, therefore painting should be blocked while
+  // zooming in/out, and zooming in/out should be block white painting
+  // TODO when the canvas is in play mode, any interaction should deactivate it
   return (
-    <Container onWheel={onWheel}>
+    <HeadlessPanel onWheel={onWheel}>
       <Background {...stats} {...canvas} />
-      <FrameLayers data-id="layers-below" layers={layersBelow} {...canvas} />
-      <FrameLayers
-        data-id="current-layer"
-        ref={setMainRef}
-        layers={[layer]}
-        {...canvas}
-      />
-      <FrameLayers data-id="layers-above" layers={layersAbove} {...canvas} />
-      <CanvasLayer ref={setPreviewRef} {...canvas} />
+      {isPlaying ? (
+        <Frames isPlaying={isPlaying} {...canvas} />
+      ) : (
+        <>
+          <FrameLayers
+            data-id="layers-below"
+            layers={layersBelow}
+            {...canvas}
+          />
+          <FrameLayers
+            data-id="current-layer"
+            ref={setMainRef}
+            layers={[layer]}
+            {...canvas}
+          />
+          <FrameLayers
+            data-id="layers-above"
+            layers={layersAbove}
+            {...canvas}
+          />
+          <CanvasLayer ref={setPreviewRef} {...canvas} />
+        </>
+      )}
       <Mask {...canvas} />
       <Float>
         <PanelSelect />
+        {button}
+        <CenterButton onClick={onCenter} />
       </Float>
-    </Container>
+      <Tools />
+    </HeadlessPanel>
   );
 };