+++ src/reducers/canvasReducer.ts
@@ -1,6 +1,35 @@
-import { getScaleAndPosition } from '../../utils';
-import { CanvasActions, actionType, Action, CanvasUpdate } from './types';
-import { Size, Canvas, Dispatch } from '../../types';
+import { getScaleAndPosition } from '../utils';
+import { Size, Canvas, Dispatch, Sprite } from '../types';
+
+export interface CanvasUpdate {
+  x?: number;
+  y?: number;
+  scale?: number;
+}
+
+export enum actionType {
+  CENTER,
+  UPDATE,
+}
+
+export interface Stats {
+  width: number;
+  height: number;
+  top: number;
+  left: number;
+}
+
+export interface CanvasActions {
+  center: (stats: Stats, sprite: Sprite) => void;
+  update: (payload: Canvas) => void;
+}
+
+type Payload = CanvasUpdate;
+
+export interface Action {
+  type: actionType;
+  payload: Payload;
+}
 
 const reducer = (state: Canvas, action: Action): Canvas => {
   const { type, payload } = action;